+++ src/main/java/selection/LoadListener.java
@@ -58,13 +58,13 @@ public class LoadListener implements ActionListener
 					return;
 				}
 				
-				final Point p1 = new Point( new float[] { Float.parseFloat( l[ 0 ] ), Float.parseFloat( l[ 1 ] ), Float.parseFloat( l[ 2 ] ) }, 
-											new float[] { Float.parseFloat( l[ 3 ] ), Float.parseFloat( l[ 4 ] ), Float.parseFloat( l[ 5 ] ) } );
+				final Point p1 = new Point( new double[] { Double.parseDouble( l[ 0 ] ), Double.parseDouble( l[ 1 ] ), Double.parseDouble( l[ 2 ] ) }, 
+											new double[] { Double.parseDouble( l[ 3 ] ), Double.parseDouble( l[ 4 ] ), Double.parseDouble( l[ 5 ] ) } );
 				
-				final Point p2 = new Point( new float[] { Float.parseFloat( l[ 8 ] ), Float.parseFloat( l[ 9 ] ), Float.parseFloat( l[ 10 ] ) }, 
-											new float[] { Float.parseFloat( l[ 11 ] ), Float.parseFloat( l[ 12 ] ), Float.parseFloat( l[ 13 ] ) } );
+				final Point p2 = new Point( new double[] { Double.parseDouble( l[ 8 ] ), Double.parseDouble( l[ 9 ] ), Double.parseDouble( l[ 10 ] ) }, 
+											new double[] { Double.parseDouble( l[ 11 ] ), Double.parseDouble( l[ 12 ] ), Double.parseDouble( l[ 13 ] ) } );
 
-				final ExtendedPointMatch pm = new ExtendedPointMatch( p1, p2, Float.parseFloat( l[ 6 ] ), Float.parseFloat( l[ 7 ] ), Float.parseFloat( l[ 14 ] ), Float.parseFloat( l[ 15 ] ) );
+				final ExtendedPointMatch pm = new ExtendedPointMatch( p1, p2, Double.parseDouble( l[ 6 ] ), Double.parseDouble( l[ 7 ] ), Double.parseDouble( l[ 14 ] ), Double.parseDouble( l[ 15 ] ) );
 				
 				matches.add( pm );
 				list1.add( Select_Points.getStringForPoint( pm.getP1(), pm.radius1W ) );
